Implemented `_DyldCacheFileRepresentable`
